pyproject.toml
[project]
name = "gis-python-tutorial"
version = "1.0.0"
requires-python = "~=3.11"
dependencies = [
    "jupyter-book>=1.0.3",
]

[project.optional-dependencies]
demo = [
    "branca>=0.8.0",
    "cartopy>=0.24.1",
    "contextily>=1.6.2",
    "folium>=0.18.0",
    "geopandas>=1.0.1",
    "ipywidgets>=8.1.5",
    "jupyterlab-widgets>=3.0.13",
    "jupyterlab>=4.2.5",
    "mapclassify>=2.8.1",
    "matplotlib>=3.8.4",
    "pooch>=1.8.2",
    "python-dotenv>=1.0.1",
    "salem>=0.3.11",
    "scikit-image>=0.24.0",
    "seaborn>=0.13.2",
    "tqdm>=4.66.6",
    "wrf-python",
]

[tool.uv.sources]
wrf-python = { git = "https://github.com/NCAR/wrf-python", branch = "develop" }

[tool.ruff]
target-version = "py311"

[tool.basedpyright]
pythonVersion = "3.11"
typeCheckingMode = "standard"

